fuck memory leak

+++ Whale/Platform/Whale/WGenericWindow.hpp
@@ -49,7 +49,7 @@ namespace Whale
 		virtual void ShowAndUpdate() = 0;
 		
 		///
-		virtual void Tick(Float deltaTime) = 0;
+		virtual void Tick(Double deltaTime) = 0;
 		
 		/// 开启文件拖放
 		virtual Bool EnableFileDrag() = 0;
